requirements.txt
tensorflow
numpy
matplotlib
streamlit
scipy
Pillow